Use text emails for auth token notifications (#2792)
#2669 added email notifications for expiring authentication tokens.
Those email notifications are sent as HTML emails when they're meant to
be sent as plaintext emails, resulting in poor formatting in most modern
email clients. We'd like to eventually move everything to HTML emails,
but that's a future endeavor. For now, this PR issues the auth token
expiration emails as plaintext.