Remove integration with non-git repos (#2795)
This PR removes all but GitHub and GitLab from the list of repositories
CDash can integrate with. This is the first in a series of PRs reducing
our integration with lesser-used external tools, paving the way for
better integration with git-based tools in the future. None of the code
removed in this PR is tested, and it is quite likely that some of these
no longer work in the first place.