Delete web-based dependency audit/upgrade (#2799)
CDash provides the ability to update JS and PHP dependencies via the UI.
Beyond the danger of potentially updating dependencies to incompatible
versions, this feature is rather silly in containerized deployments
because the changes are not baked into the image. Removing this feature
paves the way for us to have a read-only source tree. Additionally, this
change means that we can ship images without nodejs/npm or composer in
the final layer at all.