Remove obsolete `userStatistics.php` page (#2796)
`userStatistics.php` no longer provides meaningful information which
can't otherwise be obtained from git hosting platforms. The underlying
logic is also buggy, slow, and in desperate need of a refactor. This PR
removes `userStatistics.php` and all associated information. If there is
a need for a similar feature in the future, it can be re-implemented at
that time.