Lock APP_URL in browser and cypress tests (#2812)
All of our browser tests currently require exclusivity because they
change the `.env`. It isn't actually necessary to run serially if we can
guarantee that the only thing being changed in the `.env` is the
`APP_URL`. This PR resolves the issue by adding a new `RESOURCE_LOCK`
and changing the existing browser tests to allow parallel execution.
This is especially important going forward because the browser tests are
some of the longest-running tests we have.