Remove `upgrade.php` (#2816)
In a world where most users deploy CDash in a containerized environment,
`upgrade.php` has become less and less relevant, with most of the
remaining useful functionality getting moved to Artisan commands. The
remaining functionality is either broken, obsolete, or easily replaced
with more effective automations. This PR deletes the rest of the page.