Remove support for automatic mailing list user creation (#2783)
CDash currently supports the ability to create users at submission time,
a feature which is incompatible with the registration and project
membership controls required by most CDash users today. While the
feature in its current iteration is no longer relevant, it's possible
that a variation of this feature which doesn't blindly create users
could be brought back at a future point in time.