Restrict user email access (#2818)
Users' email addresses are currently considered public information. This
PR changes that, instead limiting access to users themselves and CDash
administrators. This opens the possibility for a user-specific
configuration option to allow users to optionally display their email
address in various places in the UI. As part of this PR, new `me`,
`user(id)`, and `users` fields have been added to the GraphQL schema.