docker: stop listening on port 443 (#2829)
CDash's Debian-based Docker image runs as a non-root user (www-data) yet
attempts to bind a privileged port (443) because we enable the ssl
module for production builds. This causes "Permission denied" errors in
some cases.

Change this configuration to only listen on port 8080 (which is where
CDash's traffic is served by default anyway).