Increase robustness of `TestOutput::DecompressOutput()`
We currently base64 encode test output for Postgres, but not MySQL.
This difference in behavior is a problem as we plan to drop suport for
MySQL in CDash 4.0. Migrating data from MySQL to Postgres (using pgloader)
results in a Postgres database that contains test output which is not
base64 encoded. This currently results in an error when attempting to view
this data on CDash.

Update `TestOutput::DecompressOutput()` to mitigate this issue by allowing
Postgres-backed CDash instances to decompress data that originally came from
a MySQL database, and thus isn't base64 encoded.