Limit scope of default route handler (#2831)
The current default route handler was initially conceived as a way to
integrate Laravel into the existing routing system. This route handler
only applies to a handful of legacy API routes at this point, falling
through with a 404 otherwise. This PR limits the scope of the default
route handler to only apply to routes under `/api`, using Laravel's
default route handler for the rest of the routes. I also changed the
visibility of most of the functions in the file to better reflect their
purpose.