Remove `wget` from Debian-based image (#2839)
`wget` is only used to download `composer` as part of the Debian-based
image build process. The version of `wget` embedded in a number of
recent CDash releases is vulnerable to
[CVE-2024-38428](https://nvd.nist.gov/vuln/detail/CVE-2024-38428),
causing security scanners to flag our images unnecessarily. We already
use `curl` for CDash itself, so we can simply replace all existing
usages of `wget` with `curl` to resolve this issue.