Add admin button to delete project members (#2856)
Following the work in https://github.com/Kitware/CDash/pull/2778, this
PR adds an admin-only column containing a button which removes project
members to the project members page. As of now, anyone with project
admin access can delete members. I plan to make a follow-up PR which
fine-tunes this behavior when 3rd-party authentication providers are
used.