Disable manual project membership when LDAP enabled (#2870)
It doesn't make sense to ever give users the ability to add or remove
users to a project if a project's "LDAP filter" is set. Even though it's
currently possible to add and remove users manually when an LDAP filter
is set, the changes will be overwritten automatically by the LDAP sync
script. This PR explicitly disallows manual membership changes by users
of any access level, including global administrators.