Bump `laravel-graphiql` to 3.3.3 (#2878)
As described in https://github.com/Kitware/CDash/pull/2876, the version
of `laravel-graphiql` we use contains unpinned dependencies which
currently lead to GraphiQL being broken on all versions of CDash.
https://github.com/Kitware/CDash/pull/2876 implemented a long-term fix
which will be released in CDash 4.0, but many users are likely to remain
on 3.10 for some time while transitioning from MySQL to Postgres. This
PR makes a similar change in the 3.10 release series to get GraphiQL
working again in CDash 3.x.