Fix project and global invitation access control issues (#2894)
https://github.com/Kitware/CDash/pull/2778 and
https://github.com/Kitware/CDash/pull/2830 both suffer from a logical
flaw which was missed in testing because the Laravel `assertJson(...,
strict: true)` does not actually do an exact JSON match as it was
believed to do. This PR resolves the access control issues in these
unreleased features. I plan to make a separate PR to convert the rest of
the `assertJson()` assertions throughout the codebase to
`assertExactJson()`.