Move coverage diff to `coveragesummary` table (#2896)
The coverage-related portion of our database schema is currently pretty
messy. This PR aims to clean up some of it by moving the contents of the
`coveragesummarydiff` table to the `coveragesummary` table. Eventually,
the contents of `coveragesummary` (and what was previously in
`coveragesummarydiff`) should probably be moved to the `build` table
like the rest of our aggregate summary information.