Make partial text match filters case-insensitive (#2893)
Some users previously used MySQL's case-insensitivity to deliberately
create case-insensitive filters for build and test names. While it's not
practical to implement case-insensitivity for very selective filters
while retaining their current performance, there is no real difference
in performance for filters which are not selective enough to make use of
indexes. This PR changes all of the "LIKE"-based filters to use "ILIKE"
instead. Note that this will make the "starts with" filter less
performant, in exchange for maintaining consistency with the other
LIKE-based filters.