Expose coverage file path via GraphQL (#2926)
The filename for coverage files is stored in the `coveragefile` table.
While it should eventually be moved to the `coverage` table, doing so is
nontrivial. This PR exposes it via the `Coverage` type, with the caveat
that it cannot be used in filters until the column is fully moved to the
`coverage` table.