Clean up legacy test dependencies (#2925)
This PR supersedes https://github.com/Kitware/CDash/pull/2903, moving
tests which have no dependencies to the parallel section at the
beginning of the test suite instead of creating parallel blocks in the
middle of the test suite. This approach is far safer than the aggressive
approach taken in https://github.com/Kitware/CDash/pull/2903, reducing
the chances of additional flakiness being introduced. In addition to
improving the overall test time, these changes make it easier to debug
tests late in the test by reducing the length of the longest dependency
chain.