Specify multiple processors for resource-intensive tests (#2935)
The first part of our test suite consists of a number of tests which can
be run up to a relatively high parallel level. Some of these tests
require more than one processor, which causes thrashing when multiple
tests are run concurrently. This PR attempts to address the issue by
specifying 2 processors for PHPStan, Cypress, and Dusk tests.