Display PHPUnit deprecation warnings in log (#2939)
PHPUnit records the number of deprecations in the summary by default,
without displaying the text of the actual deprecation warning. While
this makes sense for users running PHPUnit directly, it's easy to miss
in our test suite because it's embedded within the output of each test.
This PR makes the deprecation warnings more obvious by printing the text
to the test log.