Decompress test output (#2941)
CDash optionally stores test output information in a compressed format.
While it does save a small amount of storage space, this approach is
problematic for a multitude of reasons, including the fact that the data
cannot be queried by the database itself. It's also unnecessary because
[TOAST](https://www.postgresql.org/docs/current/storage-toast.html
handles compression automatically. This PR removes the compression logic
and adds a migration to decompress the entire `testoutput` table. Given
that this table is one of the largest tables for most CDash instances,
CDash administrators should expect this migration to take a significant
amount of time.