Move submission handlers to PSR-4-compliant paths (#2988)
CDash currently uses a custom autoloader to integrate Laravel with the
legacy portion of the codebase. This means that Composer cannot generate
classmaps for a significant portion of the codebase. This PR moves all
of the XML handlers to a PSR-4-compliant namespace and corresponding
directory structure for better organization and performance.