Convert `coveragefilelog` `log` column to text (#2994)
The `coveragefilelog` `log` column stores text strings. Even if it's not
immediately human readable, the underlying data is still text. Storing
text data in a binary column type is an over-engineered abuse of
Postgres' type system. This PR changes the underlying column type to
`text` for better compatibility with various parts of our PHP
application layer.