Allow coverage file path filtering (#2996)
https://github.com/Kitware/CDash/pull/2926 used a hack to allow file
paths to be accessed via the `Coverage` type, despite residing in a
separate relation. This PR re-works that to instead use a Postgres view
to access all coverage information in a single place. Subsequent PRs
will increase usage of this view throughout the codebase, with the
ultimate goal of managing the underlying tables entirely within
Postgres.