Introduce service infrastructure (#3007)
It's common for web apps built with Laravel or similar frameworks to use
"services" as wrappers around CRUD operations on models when additional
logic is needed, such as when relationships need to be set up. CDash
currently has a few service-like classes such as `TestCreator`, as well
as s substantial amount of duplicated logic which should be moved to a
common wrapper. This PR introduces an `AbstractService` base class and a
`ProjectService` which is used to initialize build groups at project
setup time.