Use strict equality for string constant comparisons (#3050)
Changes from non-strict to strict equality which are only tangentially
related to a PR are a common cause of conflicts. PHPStan baseline
conflicts are also common. Changing to strict equality when one operand
is a constant string of nonzero length is guaranteed to be safe. This PR
converts all such cases to strict equality purely for the sake of
reducing the number of errors of this type in the codebase in the hopes
of reducing future conflicts.