Remove label subscriptions (#3066)
CDash historically offered the ability to subscribe to notifications on
a per-label basis, meaning that users could subscribe to the portion of
the errors relevant to them without being bothered by errors related to
other parts of the codebase. This functionality is currently broken,
causes performance issues for several fundamental reasons, and is
entirely untested. This PR removes the feature entirely. If a similar
feature is desired in the future, we can re-evaluate it as a new feature
designed and written from the ground up.