Make saml2 proxyVars configurable (#3074)
When running CDash behind a load balancer that terminates TLS, SAML
authentication fails because there is a protocol mismatch (http vs.
https). Setting the config/saml2.php proxyVars variable to true makes it
trust the X-Forwarded-Proto header, but there is no way to configure
this via the environment.

This allows to set this parameter via the SAML2_PROXY_VARS environment
variable.