Change internal links from `/build/*` to `/builds/*` (#3084)
Many of the links throughout the site haven't been updated to reflect
the plural resource naming scheme, instead relying on redirects. This PR
changes all of the legacy `/build/*` links to instead point to
`/builds/*` routes. I also added a redirect for the various files pages
to bring them in line with the naming conventions used elsewhere.