Index `build2test(buildid,status)` (#3117)
There's currently no way for Postgres to efficiently query tests by
status for a given build. The best query plan with the current setup is
simply for Postgres to scan all of the tests for a given build. Scanning
all tests is reasonably efficient in most cases, but the overhead adds
up over time for this extremely common use case. This PR indexes
`build2test(buildid,status)` to allow the database to count the number
of passing or failing tests directly.