Move build pruning to a scheduled job (#3131)
Builds are currently pruned on the first submit of the day. This PR
moves build pruning out of the "daily update" process and into the new
scheduled jobs framework. This reduces the overhead for each submission
and cleans up what has been a major source of problems over the years.
The scheduled job runs daily. In the future, I think we should evaluate
whether the `autoremovemaxbuilds` project field is useful. Removing it
would allow us to run this job hourly.