Update links to redirect routes throughout site (#3127)
We've pluralized many routes over the last few years, but links to some
of the old routes remain across the site. These links to redirect routes
cause slightly more overhead time when clicking around the site and
increase server load. This PR addresses most of the links to redirect
routes throughout the site.