Fix flaky `sub-project-group-order` e2e test (#3136)
The `sub-project-group-order` test has been flaky for years due to a
rendering race condition in which Cypress selects the first `<a>` tag
available on the page, something which changes depending on the page
load speed. This PR fixes the issue by using a more specific selector to
precisely target the tab in question.

Example failure here:
https://github.com/Kitware/CDash/actions/runs/17947517456/job/51041305899