Sort testSummary rows deterministically in e2e test (#3135)
`test-summary.cy.js` has been flaky for years and is quite a nuisance
when unrelated changes modify the test timing subtly and upset this
delicate test. See the trouble merging
https://github.com/Kitware/CDash/pull/3132 for an example. My hypothesis
is that somehow the rows in the table are not rendered
deterministically. This PR attempts to resolve that by sorting by build
stamp. I intend to run the tests 10 times before opening this for review
to ensure that this fix actually does something.