Add test history calendar view (#3139)
When viewing a particular test, it's currently possible to view prior
test statuses on a scatter plot. A scatter plot with binary is not ideal
for a multitude of reasons, so this PR re-implements it to instead show
the data on a calendar plot. Days with only passing tests are shown in
free, days with only failing tests are shown in red, and days with both
passing and failing tests are shown in green/red. Clicking on a day
takes users to the test query page with appropriate filters set for
further exploration. Further refactoring of this plot and similar plots
is yet to come, but these changes are sufficient in the short term. I
also intend to add a visual difference test for this component once I
implement visual difference testing.

Before:
<img width="500" alt="image"
src="https://github.com/user-attachments/assets/64b2e045-85b3-4494-9978-80db0761683b"
/>


After:
<img width="500" alt="image"
src="https://github.com/user-attachments/assets/68113d2a-48ee-4e33-a3ed-83a944182c2b"
/>