Move all remaining daily updates code to scheduled job (#3173)
The "daily updates" process is currently activated on every submission,
checking the database to see if each submission is the first of the day
for a given project. This PR moves all of the logic to a scheduled daily
job, eliminating the need to perform expensive queries on every
submission.