Simplify submission processing DB queries (#3174)
The submission handling process is currently a delicate mess which is
difficult to change without breaking unrelated features. This PR is part
one of a cleanup of the submission process, using Eloquent wherever
possible and eliminating unnecessary DB queries. I'll perform a larger
cleanup in a separate PR once this preliminary work is merged.