Add reusable code box UI component (#3203)
Building on the work in #3108, this PR adds a reusable code box with
line numbers for use across the site. This component is expected to
eventually replace all of the existing prepared text boxes used for
displaying build and test output. I used the new component on the build
notes page as an example and will create follow-up PRs to increase the
usage across the site.