Show tests for child builds on new build tests page (#3205)
The legacy `viewTest.php` page shows all tests for the input build ID,
as well as tests associated with children of that build. This is an
important capability for projects with subprojects, and adding this to
the "new" build tests page is a critical step before it can replace
`viewTest.php`. To do this, I was forced to get rid of pagination, but
load testing showed that the pagination was not particularly useful in
the first place.