Move contents of `subproject2build` table to `build` (#3219)
Builds are associated with at most one subproject, but there is
currently a `subproject2build` pivot table in the schema. This PR
removes the pivot table and moves the `subprojectid` column to the
`build` table. In addition to reducing complexity, this change also
makes lookups of builds by subproject more efficient.