Add subproject to new `/builds/<id>/tests` page (#3221)
The last missing piece of the replacement page for `viewTest.php` is a
subproject column. This PR adds that column and exposes the most minimal
amount of SubProject data necessary to support it via the GraphQL API. A
follow-up PR will switch all `viewTest.php` references to use this new
page instead.