Remove unused build properties page (#3222)
`buildProperties.php` is not linked to anywhere, and is almost entirely
broken. Additionally, the page would need a complete rewrite to switch
from AngularJS to Vue in the future anyway. This PR deletes the page and
deprecates the API endpoint. Note that the concept of build properties
is not going away entirely: build properties are still used on several
other pages. It's just the dedicated build properties page that's being
removed.