Clean up submission handler undefined fields and types (#3228)
This maintenance PR achieves three goals:
1. Explicitly define all fields on every submission handler.
(dynamically defining fields is deprecated, and will be removed in a
future version of PHP)
2. Move field initialization out of the constructor wherever possible
3. Add type annotations to fields where it's easy to do so

Note that (3) does not attempt to add type annotations to all fields,
but rather just those for which the type is easily identified and safe
to add. A follow-up PR will be created to handle the more difficult
cases. Adding type annotations makes refactoring other parts of the
codebase safer because static analysis tools have a better depth of
analysis.