Add instrumentation links to build summary page (#3229)
By adding links to `/builds/<id>/targets` and `/builds/<id>/commands` to
the build summary page, these instrumentation pages are now official
CDash features users can navigate to though normal usage of the website.
I intend to improve the integration of these pages with the rest of the
site as part of a future revamp of the build summary page.