Move legacy project relationship methods to project service (#3293)
Historically, models provided helper methods to access relationships for
specific needs. Custom helpers like this are an antipattern in active
record systems like Eloquent, so it doesn't make sense to simply move
these helpers over to our Eloquent models. Instead, it's traditional for
services to fill this role. As such, this PR moves the methods which
only depend on a model's project ID to the `ProjectService` class. With
much of the business logic in the legacy `Project` model out of the way,
this should enable us to migrate away from the legacy model more
quickly.