Restore notes link on build summary page (#3299)
https://github.com/Kitware/CDash/pull/3259 updated the header of the
build summary page, which deleted the link to the notes page in the
process. This PR re-adds it as a dedicated section. I plan to move all
of the links to `/builds/<id>/*` pages to a new sidebar UI element in
the next 1-3 releases, so I expect this solution to be a temporary fix.