Migrate legacy AngularJS code to ESM (#3307)
Historically, CDash loaded all JS into the global namespace, including
all of our AngularJS and older code. That approach makes tracing
dependencies a nightmare and makes it essentially impossible to allow
Vue and AngularJS to coexist on a single page. This PR resolves most of
the issue by creating a single `legacy.js` file containing everything
not already contained within the Vue ecosystem. The one exception is
jQuery, which must be in the window namespace for our legacy jQuery
plugins to function. This PR paves the way for a follow-up PR switching
from Mix/Webpack to Vite.